Honor PUBLIC_URL in development
